+++ changelog.md
@@ -7,6 +7,7 @@ v0.6.2.4 June 2019
 v0.6.2.3 October 2018
 
  * GHC 8.6 fixes
+ * Adding: Allow definitions to appear anywhere in an svg document.
 
 v0.6.2.2 December 2017